apply key to all entities
