+++ src/packages/core/src/federation/index.ts
@@ -3,7 +3,6 @@ import { addEntitiesQuery } from './entities';
 import { addEnums } from './enums';
 import { addServiceQuery } from './service';
 
-// enable federation
 export const enableFederation = () => {
 	addEnums();
 	addDirectives();
